Add admin course group navigation and reorganize pages
