+++ Pages/Admin/CourseGroups/Create.cshtml.cs
@@ -1,18 +1,17 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
 using SysJaky_N.Data;
 using SysJaky_N.Models;
 
-namespace SysJaky_N.Pages.CourseGroups;
+namespace SysJaky_N.Pages.Admin.CourseGroups;
 
 [Authorize(Policy = AuthorizationPolicies.AdminOnly)]
-public class EditModel : PageModel
+public class CreateModel : PageModel
 {
     private readonly ApplicationDbContext _context;
 
-    public EditModel(ApplicationDbContext context)
+    public CreateModel(ApplicationDbContext context)
     {
         _context = context;
     }
@@ -20,15 +19,8 @@ public class EditModel : PageModel
     [BindProperty]
     public CourseGroup CourseGroup { get; set; } = new();
 
-    public async Task<IActionResult> OnGetAsync(int id)
+    public void OnGet()
     {
-        var group = await _context.CourseGroups.FindAsync(id);
-        if (group == null)
-        {
-            return NotFound();
-        }
-        CourseGroup = group;
-        return Page();
     }
 
     public async Task<IActionResult> OnPostAsync()
@@ -38,7 +30,7 @@ public class EditModel : PageModel
             return Page();
         }
 
-        _context.Attach(CourseGroup).State = EntityState.Modified;
+        _context.CourseGroups.Add(CourseGroup);
         await _context.SaveChangesAsync();
         return RedirectToPage("Index");
     }
