+++ Pages/Admin/CourseGroups/Delete.cshtml
@@ -1,5 +1,5 @@
 @page "{id:int}"
-@model SysJaky_N.Pages.CourseGroups.DeleteModel
+@model SysJaky_N.Pages.Admin.CourseGroups.DeleteModel
 @{
     ViewData["Title"] = "Delete Group";
 }
