+++ Pages/Admin/CourseGroups/Edit.cshtml
@@ -1,18 +1,19 @@
-@page
-@model SysJaky_N.Pages.CourseGroups.CreateModel
+@page "{id:int}"
+@model SysJaky_N.Pages.Admin.CourseGroups.EditModel
 @{
-    ViewData["Title"] = "Create Group";
+    ViewData["Title"] = "Edit Group";
 }
 
-<h1>Create Group</h1>
+<h1>Edit Group</h1>
 
 <form method="post">
+    <input type="hidden" asp-for="CourseGroup.Id" />
     <div class="form-group">
         <label asp-for="CourseGroup.Name"></label>
         <input asp-for="CourseGroup.Name" class="form-control" />
         <span asp-validation-for="CourseGroup.Name" class="text-danger"></span>
     </div>
-    <button type="submit" class="btn btn-primary">Create</button>
+    <button type="submit" class="btn btn-primary">Save</button>
     <a asp-page="Index" class="btn btn-secondary">Back to List</a>
 </form>
