+++ Pages/Admin/CourseGroups/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model SysJaky_N.Pages.CourseGroups.IndexModel
+@model SysJaky_N.Pages.Admin.CourseGroups.IndexModel
 @{
     ViewData["Title"] = "Course Groups";
 }
