+++ Pages/Admin/CourseGroups/Index.cshtml.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using SysJaky_N.Data;
 using SysJaky_N.Models;
 
-namespace SysJaky_N.Pages.CourseGroups;
+namespace SysJaky_N.Pages.Admin.CourseGroups;
 
 [Authorize(Policy = AuthorizationPolicies.AdminOnly)]
 public class IndexModel : PageModel
