+++ Pages/Admin/_Layout.cshtml
@@ -33,6 +33,9 @@
                             <a class="nav-link @ActiveClass("/Admin/Courses")" asp-page="/Admin/Courses/Index">Courses</a>
                         </li>
                         <li class="nav-item" role="presentation">
+                            <a class="nav-link @ActiveClass("/Admin/CourseGroups")" asp-page="/Admin/CourseGroups/Index">Skupiny kurzů</a>
+                        </li>
+                        <li class="nav-item" role="presentation">
                             <a class="nav-link @ActiveClass("/Admin/CourseTerms")" asp-page="/Admin/CourseTerms/Index">Course Terms</a>
                         </li>
                         <li class="nav-item" role="presentation">
